Add ability for controllers to register child controllers.
This will allow dynamic controller trees to be created. Controllers
can register other controllers from their `handler` function. If the
controller is registered with the static params, then the child controller
will be started immediatelly. Otherwise, child controller will be
normally managed by the controller manager - started and stopped based
on the route params.

When the parent controller is stopped, child
controllers are stopped and deregistered (potentially this will include
the whole subtree). Calling `register-child-controller` with the different
static params than previously will stop the previous version of the controller
and start the new one.

API:

- `register-child-controller` - Registers and potentially starts the child controller
- `register-child-controllers` - Registers multiple child controllers
- `deregister-child-controller` - Stops the child controller if it was started and deregisters it
- `deregister-child-controllers` - Deregisters multiple child controllers
- `synchronize-child-controllers` - Registers and deregisters controllers as needed - based on the params
- `get-children` - Returns child controller names and their params
- `get-parent` - Returns name of the parent controller